cost should be long

+++ parallel1/mpi_ant_colony.cpp
@@ -33,8 +33,8 @@ int main(int argc, char* argv[]) {
   int *bestPath;
   int *otherBestPath;
   int *currentPath;
-  int bestCost = INFTY;
-  int otherBestCost;
+  long bestCost = INFTY;
+  long otherBestCost;
   float* localPheromonsPath;
   float* otherPheromonsPath;
 
@@ -306,7 +306,7 @@ int main(int argc, char* argv[]) {
         }
 
         // update bestCost and bestPath
-        int oldCost = bestCost;
+        long oldCost = bestCost;
         bestCost = updateBestPath(bestCost, bestPath, currentPath, map, nCities);
 
         if (oldCost > bestCost) {
@@ -336,7 +336,7 @@ int main(int argc, char* argv[]) {
     for (j = 0; j < nCities*nCities; j++) {
       pheromonsUpdate[j] = 1;
     }
-    int tempBestCost = bestCost;
+    long tempBestCost = bestCost;
     int* tempBestPath = (int*) malloc(nCities * sizeof(int));
     int tempTerminationCondition = terminationCondition;
     copyVectorInt(bestPath, tempBestPath, nCities);
@@ -362,7 +362,7 @@ int main(int argc, char* argv[]) {
         MPI_Finalize();
         return -1;
       }
-      if (MPI_Bcast(&otherBestCost, 1, MPI_INT, i, MPI_COMM_WORLD) != MPI_SUCCESS) {
+      if (MPI_Bcast(&otherBestCost, 1, MPI_LONG, i, MPI_COMM_WORLD) != MPI_SUCCESS) {
         printf("Node %d : Error in Broadcast of otherTerminationCondition", prank);
         MPI_Finalize();
         return -1;
@@ -413,7 +413,7 @@ int main(int argc, char* argv[]) {
         MPI_Finalize();
         return -1;
       }
-      int oldCost = bestCost;
+      long oldCost = bestCost;
       bestCost = updateBestPath(bestCost, bestPath, otherBestPath, map, nCities);
 
       if (oldCost > bestCost) {
@@ -431,7 +431,7 @@ int main(int argc, char* argv[]) {
 
   if (prank == 0) {
     // printPath(bestPath, nCities);
-    printf("best cost : %d\n", bestCost);
+    printf("best cost : %ld\n", bestCost);
   }
 
   if (prank == 0) {
